refactor: create objects with methods
