+++ src/js/components/common.js
@@ -1,0 +1,9 @@
+export const common = {
+    getFolders() {
+        return document.querySelectorAll('.folder')
+    },
+
+    getFiles() {
+        return document.querySelectorAll('.file')
+    },
+}
